+++ app/minime/forms.py
@@ -1,0 +1,13 @@
+from django import forms
+
+
+class ShortenURLForm(forms.Form):
+    url = forms.URLField(
+        required=True,
+        label='URL',
+        widget=forms.URLInput(attrs={
+            'placeholder': 'URL to shorten',
+            'autofocus': True,
+            'class': 'form-control',
+        })
+    )