+++ app/minime/models.py
@@ -1,10 +1,9 @@
-# -*- coding: utf-8 -*-
-
 from django.db import models
 
+
 class Urls(models.Model):
-    short_id = models.SlugField(max_length=6,primary_key=True)
-    httpurl = models.URLField(max_length=200)
+    short_id = models.SlugField(max_length=6, primary_key=True)
+    httpurl = models.URLField(max_length=2048)
     pub_date = models.DateTimeField(auto_now=True)
     count = models.IntegerField(default=0)