+++ app/minime/views.py
@@ -1,45 +1,51 @@
-# -*- coding: utf-8 -*-
 
-import random, string, json
+import json
+import random
+import string
 
-from django.shortcuts import render_to_response, get_object_or_404, render
-from shortenerSite.models import Urls
-from django.http import HttpResponseRedirect, HttpResponse
 from django.conf import settings
-from django.core.context_processors import csrf
+from django.http import HttpResponse, HttpResponseRedirect
+from django.shortcuts import get_object_or_404, render
+
+from .forms import ShortenURLForm
+from .models import Urls
 
 
 def index(request):
-    c = {}
-    c.update(csrf(request))
-    return render_to_response('shortenerSite/index.html', c)
+    return render(request, 'index.html', {'form': ShortenURLForm})
 
-def redirectOriginal(request, short_id):
-    url = get_object_or_404(Urls, pk=short_id) # get object, if not found return 404 error
+
+def redirect(request, short_id):
+    url = get_object_or_404(Urls, pk=short_id)  # gets object, if not found returns 404 error
     url.count += 1
     url.save()
+
     return HttpResponseRedirect(url.httpurl)
 
-def shortenUrl(request):
-    url = request.POST.get("url", '')
-    if not (url == ''):
-        short_id = getShortCode()
+
+def shorten_url(request):
+
+    url = request.POST.get('url', '')
+
+    if url:
+        short_id = get_short_code()
         b = Urls(httpurl=url, short_id=short_id)
         b.save()
 
-        response_data = {}
-        response_data['url'] = settings.SITE_URL + "/" + short_id
-        return HttpResponse(json.dumps(response_data), content_type="application/json")
-    return HttpResponse(json.dumps({"error": "error occurs"}),
-             content_type="application/json")
+        response_data = dict(url='{}/{}'.format(settings.SITE_URL, short_id))
+        return HttpResponse(json.dumps(response_data), content_type='application/json')
 
-def getShortCode():
+    return HttpResponse(json.dumps({'error': 'error occurs'}), content_type='application/json')
+
+
+def get_short_code():
     length = 6
     char = string.ascii_uppercase + string.digits + string.ascii_lowercase
-    # if the randomly generated short_id is used then generate next
+
+    # Generate a new ID, until one is found that is unique
     while True:
         short_id = ''.join(random.choice(char) for x in range(length))
         try:
-            temp = Urls.objects.get(pk=short_id)
+            Urls.objects.get(pk=short_id)
         except:
             return short_id