+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## Coming up...
 
-- First release!
+### Next
+
+- Added ability to run app on Heroku.
+- Added ability to run app in Docker.
+- Use native Django forms.
+- Added LESS support.
+- MVP version.